structure ugo into commonJS
