+++ src/shared/exponent.js
@@ -1,0 +1,13 @@
+/*! 
+vovem-ugo v2.3.1 | Ugo is mathematics an open-source library for JavaScript commonJS 
+Copyright 8/6/2022, 21:10:13 PM | VovemDev | MIT license 
+*/
+
+// Exponential values of integer
+module.exports = function exponent (a, b){
+    let e = 1;
+    for(let i = 1; i <= b; i++) {
+        e *= a;
+    }
+    return e;
+};
