+++ src/shared/procise.js
@@ -1,0 +1,9 @@
+/*! 
+vovem-ugo v2.3.1 | Ugo is mathematics an open-source library for JavaScript commonJS 
+Copyright 8/6/2022, 21:10:13 PM | VovemDev | MIT license 
+*/
+
+// To Precision integer
+module.exports = function procise (n, p){
+    return n.toPrecision(p);
+};
